+++ render_engine/src/RenderEngine.cpp
@@ -9,7 +9,7 @@ namespace RenderEngine
     {
         constexpr uint32_t kMaxNumOfResources = 10;
     }
-    RenderEngine::RenderEngine(Device& device, std::shared_ptr<CommandContext>&& command_context, size_t back_buffer_count)
+    RenderEngine::RenderEngine(Device& device, std::shared_ptr<CommandContext>&& command_context, uint32_t back_buffer_count)
         : _device(device)
         , _gpu_resource_manager(device.getPhysicalDevice(), device.getLogicalDevice(), back_buffer_count, kMaxNumOfResources)
         , _command_context(command_context->clone())
@@ -23,7 +23,7 @@ namespace RenderEngine
         VkSubmitInfo2 submit_info{};
         submit_info.sType = VK_STRUCTURE_TYPE_SUBMIT_INFO_2;
 
-        submit_info.commandBufferInfoCount = command_buffers.size();
+        submit_info.commandBufferInfoCount = static_cast<uint32_t>(command_buffers.size());
         submit_info.pCommandBufferInfos = command_buffers.data();
 
         sync_operations.fillInfo(submit_info);
