+++ render_engine/src/resources/PushConstantsUpdater.cpp
@@ -2,13 +2,13 @@
 
 namespace RenderEngine
 {
-    void PushConstantsUpdater::update(VkShaderStageFlags shader_stages, size_t offset, std::span<const uint8_t> data)
+    void PushConstantsUpdater::update(VkShaderStageFlags shader_stages, uint32_t offset, std::span<const uint8_t> data)
     {
         _logical_device->vkCmdPushConstants(_command_buffer,
                                             _layout,
                                             shader_stages,
                                             offset,
-                                            data.size(),
+                                            static_cast<uint32_t>(data.size()),
                                             data.data());
     }
 }
